initial webpack setup
