+++ package.json
@@ -7,12 +7,14 @@
   "license": "MIT",
   "scripts": {
     "serve": "live-server public/",
-    "build": "babel src/app.js --out-file=public/scripts/app.js --presets=env,react --watch"
+    "build": "webpack --watch",
+    "build-babel": "babel src/app.js --out-file=public/scripts/app.js --presets=env,react --watch"
   },
   "dependencies": {
     "babel-cli": "6.24.1",
     "babel-preset-env": "1.5.2",
     "babel-preset-react": "6.24.1",
-    "live-server": "^1.2.0"
+    "live-server": "^1.2.0",
+    "webpack": "3.1.0"
   }
 }
