+++ public/index.html
@@ -11,10 +11,11 @@
     <div id="app">
 
     </div>
-    <script src="https://unpkg.com/react@16.0.0/umd/react.development.js"></script>
-    <script src="https://unpkg.com/react-dom@16.0.0/umd/react-dom.development.js"></script>
-    <script src="/scripts/app.js">
-    </script>
+    <script src="/bundle.js"></script>
+    <!-- <script src="https://unpkg.com/react@16.0.0/umd/react.development.js"></script>
+    <script src="https://unpkg.com/react-dom@16.0.0/umd/react-dom.development.js"></script> -->
+    <!-- <script src="/scripts/app.js">
+    </script> -->
   </body>
 
 </html>
