+++ webpack.config.js
@@ -1,0 +1,12 @@
+// entry point is src/ap.js
+// output for bundle
+
+const path = require('path');
+
+module.exports = {
+  entry: './src/app.js',
+  output: {
+    path: path.join(__dirname, 'public'),
+    filename: 'bundle.js'
+  }
+};
